FINISHED

+++ src/include/Rectangle.h
@@ -2,6 +2,7 @@
 #define RECTANGLE_H
 
 #include "GenericTraits.h"
+#include <stdint.h>
 
 /// Use `DECLARE_STRUCT` to declare the struct in less than 5 lines.
 /* Your code here to declare the struct */
@@ -9,8 +10,8 @@
 DECLARE_STRUCT(
     Rectangle, {
         GenericTraits *impl;
-        int height;
-        int width;
+        uint64_t height;
+        uint64_t width;
     })
 
 #endif
